+++ bin/miden-cli/src/tests.rs
@@ -8,10 +8,10 @@ use std::{
 
 use assert_cmd::Command;
 use config::RpcConfig;
+use miden_cli::CliKeyStore;
 use miden_client::{
     self, Client, Felt,
     account::{AccountId, AccountStorageMode},
-    authenticator::{ClientAuthenticator, keystore::FilesystemKeyStore},
     crypto::{FeltRng, RpoRandomCoin},
     note::{
         Note, NoteAssets, NoteExecutionHint, NoteExecutionMode, NoteFile, NoteInputs, NoteMetadata,
@@ -19,7 +19,7 @@ use miden_client::{
     },
     rpc::{Endpoint, TonicRpcClient},
     store::sqlite_store::SqliteStore,
-    testing::account_id::ACCOUNT_ID_OFF_CHAIN_SENDER,
+    testing::account_id::ACCOUNT_ID_PRIVATE_SENDER,
     transaction::{OutputNote, TransactionRequestBuilder},
     utils::Serializable,
 };
@@ -107,7 +107,7 @@ async fn test_mint_with_untracked_account() {
 // ================================================================================================
 
 // Only one faucet is being created on the genesis block
-const GENESIS_ACCOUNTS_FILENAMES: [&str; 1] = ["faucet.mac"];
+const GENESIS_ACCOUNTS_FILENAMES: [&str; 1] = ["account_0.mac"];
 
 // This tests that it's possible to import the genesis accounts and interact with them. To do so it:
 //
@@ -164,7 +164,7 @@ async fn test_import_genesis_accounts_can_be_used_for_transactions() {
     // Let's try and mint
     mint_cli(
         &temp_dir,
-        &AccountId::try_from(ACCOUNT_ID_OFF_CHAIN_SENDER).unwrap().to_hex(),
+        &AccountId::try_from(ACCOUNT_ID_PRIVATE_SENDER).unwrap().to_hex(),
         &fungible_faucet_account_id,
     );
 
@@ -234,7 +234,7 @@ async fn test_cli_export_import_note() {
     consume_note_cli(&temp_dir_2, &first_basic_account_id, &[&note_to_export_id]);
 
     // Test send command
-    let mock_target_id: AccountId = AccountId::try_from(ACCOUNT_ID_OFF_CHAIN_SENDER).unwrap();
+    let mock_target_id: AccountId = AccountId::try_from(ACCOUNT_ID_PRIVATE_SENDER).unwrap();
     send_cli(
         &temp_dir_2,
         &first_basic_account_id,
@@ -637,7 +637,7 @@ pub fn create_test_store_path() -> std::path::PathBuf {
 pub type TestClient = Client<RpoRandomCoin>;
 
 /// Creates a new [`Client`] with a given store. Also returns the keystore associated with it.
-async fn create_rust_client_with_store_path(store_path: &Path) -> (TestClient, FilesystemKeyStore) {
+async fn create_rust_client_with_store_path(store_path: &Path) -> (TestClient, CliKeyStore) {
     let rpc_config = RpcConfig::default();
 
     let store = {
@@ -645,20 +645,19 @@ async fn create_rust_client_with_store_path(store_path: &Path) -> (TestClient, F
         std::sync::Arc::new(sqlite_store)
     };
 
-    let mut rng = rand::thread_rng();
-    let coin_seed: [u64; 4] = rng.r#gen();
+    let mut rng = rand::rng();
+    let coin_seed: [u64; 4] = rng.random();
 
     let rng = RpoRandomCoin::new(coin_seed.map(Felt::new));
 
-    let keystore = FilesystemKeyStore::new(temp_dir()).unwrap();
+    let keystore = CliKeyStore::new(temp_dir()).unwrap();
 
-    let authenticator = ClientAuthenticator::new(rng, keystore.clone());
     (
         TestClient::new(
             Arc::new(TonicRpcClient::new(&rpc_config.endpoint.into(), rpc_config.timeout_ms)),
             rng,
             store,
-            std::sync::Arc::new(authenticator),
+            std::sync::Arc::new(keystore.clone()),
             true,
         ),
         keystore,