+++ crates/rust-client/src/store/web_store/transaction/mod.rs
@@ -58,7 +58,7 @@ impl WebStore {
                 let output_notes = OutputNotes::read_from_bytes(&tx_idxdb.output_notes)?;
 
                 let transaction_script: Option<TransactionScript> =
-                    if tx_idxdb.script_hash.is_some() {
+                    if tx_idxdb.script_root.is_some() {
                         let tx_script = tx_idxdb
                             .tx_script
                             .map(|script| TransactionScript::read_from_bytes(&script))
@@ -70,8 +70,11 @@ impl WebStore {
                         None
                     };
 
-                let transaction_status =
-                    commit_height.map_or(TransactionStatus::Pending, TransactionStatus::Committed);
+                let transaction_status = match (commit_height, tx_idxdb.discarded) {
+                    (_, true) => TransactionStatus::Discarded,
+                    (Some(block_num), false) => TransactionStatus::Committed(block_num),
+                    (None, false) => TransactionStatus::Pending,
+                };
 
                 Ok(TransactionRecord {
                     id: id.into(),